requirements.txt
Flask==2.0.1
opencv-python==4.5.3
numpy==1.21.2
Pillow==8.2.0